requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic[email]
python-multipart
pillow
razorpay
gunicorn
cloudinary